mend

+++ src/quadtree.cpp
@@ -28,7 +28,7 @@ void QuadTree::insert(ObjectQuadTree* objectToInsert){
 
 	if(_nodesVector.empty()){
 		_objectVector.push_back(objectToInsert);
-		if(_objectVector.size() > this->_maxObjectsInQuadrant){
+		if((int)_objectVector.size() > this->_maxObjectsInQuadrant){
 			if(this->_layer < this->_maxLayers - 1){
 				this->_split();
 			}
